+++ Duende.IdentityServer.Contrib.RedisStore.Tests/appsettings.json
@@ -1,0 +1,5 @@
+{
+  "Redis": {
+    "ConnectionString": "localhost"
+  }
+}
