+++ Duende.IdentityServer.Contrib.RedisStore/Duende.IdentityServer.Contrib.RedisStore.nuspec
@@ -1,0 +1,15 @@
+<?xml version="1.0"?>
+<package >
+  <metadata>
+    <id>$id$</id>
+    <version>$version$</version>
+    <title>$title$</title>
+    <authors>Safe Fleet</authors>
+    <owners>Safe Fleet</owners>
+    <requireLicenseAcceptance>false</requireLicenseAcceptance>
+    <description>$description$</description>
+    <releaseNotes>Add support to use already established connection through ConnectionMultiplexer</releaseNotes>
+    <copyright>Copyright 2021</copyright>
+    <tags>Duende IdentityServer Redis Store</tags>
+  </metadata>
+</package>
