--- IdentityServer4.Contrib.RedisStore.Tests/appsettings.json
@@ -1,5 +1,0 @@
-﻿{
-	"Redis": {
-		"ConnectionString":  "localhost" 
-	} 
-}
