--- IdentityServer4.Contrib.RedisStore.sln
@@ -1,31 +1,0 @@
-﻿
-Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio 15
-VisualStudioVersion = 15.0.26730.8
-MinimumVisualStudioVersion = 10.0.40219.1
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "IdentityServer4.Contrib.RedisStore", "IdentityServer4.Contrib.RedisStore\IdentityServer4.Contrib.RedisStore.csproj", "{B6285ED9-1D6D-49FB-B9A7-5C5460415E86}"
-EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "IdentityServer4.Contrib.RedisStore.Tests", "IdentityServer4.Contrib.RedisStore.Tests\IdentityServer4.Contrib.RedisStore.Tests.csproj", "{B2273D4D-FBD6-4665-9856-16F08527571D}"
-EndProject
-Global
-	GlobalSection(SolutionConfigurationPlatforms) = preSolution
-		Debug|Any CPU = Debug|Any CPU
-		Release|Any CPU = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{B6285ED9-1D6D-49FB-B9A7-5C5460415E86}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{B6285ED9-1D6D-49FB-B9A7-5C5460415E86}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{B6285ED9-1D6D-49FB-B9A7-5C5460415E86}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{B6285ED9-1D6D-49FB-B9A7-5C5460415E86}.Release|Any CPU.Build.0 = Release|Any CPU
-		{B2273D4D-FBD6-4665-9856-16F08527571D}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{B2273D4D-FBD6-4665-9856-16F08527571D}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{B2273D4D-FBD6-4665-9856-16F08527571D}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{B2273D4D-FBD6-4665-9856-16F08527571D}.Release|Any CPU.Build.0 = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(SolutionProperties) = preSolution
-		HideSolutionNode = FALSE
-	EndGlobalSection
-	GlobalSection(ExtensibilityGlobals) = postSolution
-		SolutionGuid = {258B4519-0BDA-45DD-8FA5-A46D1BC27A4D}
-	EndGlobalSection
-EndGlobal
