--- IdentityServer4.Contrib.RedisStore/IdentityServer4.Contrib.RedisStore.csproj
@@ -1,51 +1,0 @@
-﻿<Project Sdk="Microsoft.NET.Sdk">
-
-  <PropertyGroup>
-    <TargetFramework>netcoreapp3.1</TargetFramework>
-    <GeneratePackageOnBuild>false</GeneratePackageOnBuild>
-    <Description>Redis Store for operational data and for Caching of Identity Server 4</Description>
-    <Authors>Ali Bazzi</Authors>
-    <Company>Ali Bazzi</Company>
-    <AssemblyVersion>4.0.0.0</AssemblyVersion>
-    <PackageReleaseNotes>Supports Identity Server 4 v4.</PackageReleaseNotes>
-    <Version>4.0.0</Version>
-    <PackageProjectUrl>https://github.com/AliBazzi/IdentityServer4.Contrib.RedisStore</PackageProjectUrl>
-    <PackageIconUrl>https://identityserver.github.io/Documentation/assets/images/icons/IDserver_icon128.jpg</PackageIconUrl>
-    <PackageIcon>icon.jpg</PackageIcon>
-    <PackageTags>Redis Store IdentityServer4</PackageTags>
-    <PackageLicenseExpression>MIT</PackageLicenseExpression>
-    <Copyright>Ali Bazzi</Copyright>
-    <FileVersion>4.0.0.0</FileVersion>
-    <PublishRepositoryUrl>true</PublishRepositoryUrl>
-    <IncludeSymbols>true</IncludeSymbols>
-    <SymbolPackageFormat>snupkg</SymbolPackageFormat>
-  </PropertyGroup>
-
-  <ItemGroup>
-    <PackageReference Include="Microsoft.SourceLink.GitHub" Version="1.0.0" PrivateAssets="All" />
-  </ItemGroup>
-
-
-  <PropertyGroup Condition="'$(Configuration)|$(Platform)'=='Debug|AnyCPU'">
-    <DocumentationFile>bin\Debug\netcoreapp3.1\IdentityServer4.Contrib.RedisStore.xml</DocumentationFile>
-  </PropertyGroup>
-
-  <PropertyGroup Condition="'$(Configuration)|$(Platform)'=='Release|AnyCPU'">
-    <DocumentationFile>bin\Release\netcoreapp3.1\IdentityServer4.Contrib.RedisStore.xml</DocumentationFile>
-  </PropertyGroup>
-
-  <ItemGroup>
-    <None Remove="IdentityServer4.Contrib.RedisStore.nuspec" />
-  </ItemGroup>
-
-  <ItemGroup>
-    <PackageReference Include="IdentityServer4" Version="4.0.0" />
-    <PackageReference Include="IdentityServer4.Storage" Version="4.0.0" />
-    <PackageReference Include="StackExchange.Redis" Version="2.1.*" />
-  </ItemGroup>
-
-  <ItemGroup>
-    <None Include="icon.jpg" Pack="true" PackagePath="\"/>
-  </ItemGroup>
-
-</Project>
