--- IdentityServer4.Contrib.RedisStore/IdentityServer4.Contrib.RedisStore.nuspec
@@ -1,15 +1,0 @@
-<?xml version="1.0"?>
-<package >
-  <metadata>
-    <id>$id$</id>
-    <version>$version$</version>
-    <title>$title$</title>
-    <authors>Ali Bazzi</authors>
-    <owners>Ali Bazzi</owners>
-    <requireLicenseAcceptance>false</requireLicenseAcceptance>
-    <description>$description$</description>
-    <releaseNotes>Add support to use already established connection through ConnectionMultiplexer</releaseNotes>
-    <copyright>Copyright 2019</copyright>
-    <tags>IdentityServer IdentityServer4 Redis Store</tags>
-  </metadata>
-</package>
